requirements.txt
apiclient==1.0.4
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
chardet==5.1.0
charset-normalizer==2.1.1
cryptography==37.0.2
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.1
django-allauth==0.51.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-excel==0.0.10
django-extensions==3.2.1
django-import-export==2.9.0
django-js-asset==2.0.0
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.33.3
google==3.0.0
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.14.2
google-cloud-core==2.3.2
google-cloud-speech==2.15.1
google-cloud-storage==2.5.0
google-cloud-texttospeech==2.12.1
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
kiwisolver==1.4.3
lml==0.1.0
MarkupPy==1.14
matplotlib==3.5.2
mssql-django==1.2
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
proto-plus==1.22.1
protobuf==4.21.5
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive2==1.14.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
PyJWT==2.6.0
pynput==1.7.6
pyodbc==4.0.35
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.5.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
texttable==1.6.7
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
